Rename razer-drivers to openrazer
Summary:
This is meant to fix one of the tickboxes in T3667
Upstream changed the name from razer-drivers to openrazer and this revision makes that change in the Solus repo too.

PR to drop razer-drivers can be found [here](https://github.com/getsolus/solus-sc/pull/9).
Revisions part of this change: D5858 and D5859

Test Plan:
- Verified that package built without errors
- Installed and verified that `razergenie` and `polychromatic` didn't print any errors or complain
(Please note that I don't have any razer hardware to test with, but this is such a small change so it should be necessary)
- Any help to just double check this would be helpful though

Reviewers: #triage_team, JoshStrobl, DataDrake

Reviewed By: #triage_team, JoshStrobl, DataDrake

Subscribers: DataDrake, JoshStrobl, Girtablulu

Differential Revision: https://dev.getsol.us/D5877